% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  fasta_filename,
  fasta_filenames,
  folder_name,
  gene_ids_filename,
  gene_names_filename,
  is_in_tmh_filename,
  is_in_tmh_filenames,
  n_gene_ids,
  n_snps,
  results_filename,
  snps_filename,
  snps_filenames,
  snps_id_filename,
  snps_id_filenames,
  topo_filename,
  topo_filenames,
  use_color,
  variations_csv_filename,
  variations_csv_filenames,
  variations_rds_filename,
  variations_rds_filenames,
  verbose
)
}
\arguments{
\item{fasta_filename}{name of a FASTA file with a protein sequence}

\item{fasta_filenames}{name of one or more FASTA files
with protein sequences}

\item{folder_name}{name of the folder}

\item{gene_ids_filename}{the filename to save
the gene IDs to.
For the experiment, use \code{gene_ids.csv}}

\item{gene_names_filename}{the filename to save
the gene IDs and gene names to,
as can be read by \link[ncbiperegrine]{read_gene_names_file}.
For the experiment, use \code{gene_names.csv}}

\item{is_in_tmh_filename}{the filename to save
the \code{is_in_tmh} table to (as can be read by
\link[ncbiperegrine]{read_is_in_tmh_file}).
For the experiment, use \code{[gene_name]_is_in_tmh.csv}}

\item{is_in_tmh_filenames}{the filename to save
the \code{is_in_tmh} tables to (each of which can be read by
\link[ncbiperegrine]{read_is_in_tmh_file}).
For the experiment, use \code{[gene_name]_is_in_tmh.csv}}

\item{n_gene_ids}{the number of gene IDs.
Use \link{Inf} to use all gene IDs}

\item{n_snps}{the number of SNPs.
Use \link{Inf} to use all SNS IDs}

\item{results_filename}{name of the results file.
For the experiment, use \code{results.csv}}

\item{snps_filename}{one filenames of a file
containing the SNP IDs,
named \code{[gene_name]_snps.csv}.
These files can be read by \link[ncbiperegrine]{read_snps_file}}

\item{snps_filenames}{one or more filenames of files
containing the SNP IDs,
named \code{[gene_name]_snps.csv}.
These files can be read by \link[ncbiperegrine]{read_snps_file}}

\item{snps_id_filename}{one filenames of a file
containing the SNP IDs,
named \code{[gene_name]_snps.csv}.
These files can be read by \link[ncbiperegrine]{read_snps_file}}

\item{snps_id_filenames}{one or more filenames of files
containing the SNP IDs,
named \code{[gene_name]_snps.csv}.
These files can be read by \link[ncbiperegrine]{read_snps_file}}

\item{topo_filename}{name of a \code{.topo} file with a protein topology}

\item{topo_filenames}{name of one or more \code{.topo} files
with protein topologies}

\item{use_color}{set to TRUE to use color,
set to FALSE for black-and-white}

\item{variations_csv_filename}{name of a \code{[gene_name]_variations.csv}
file}

\item{variations_csv_filenames}{names of one or more
\code{[gene_name]_variations.csv}
files}

\item{variations_rds_filename}{name of a \code{[gene_name]_variations.rds}
file}

\item{variations_rds_filenames}{names of one or more
\code{[gene_name]_variations.rds}
files}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
