% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_constants.R
\name{get_n_unique_protein_names_in_single_spanners}
\alias{get_n_unique_protein_names_in_single_spanners}
\title{Get this constant}
\usage{
get_n_unique_protein_names_in_single_spanners()
}
\description{
Get this constant
}
