% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mis.R
\name{is_mis}
\alias{is_mis}
\title{Determine if this is a monotonously increasing function
This is any sequence that is generated by \link{seq}
that increases in steps of one.}
\usage{
is_mis(sequence)
}
\arguments{
\item{sequence}{a sequence of numbers}
}
\description{
Determine if this is a monotonously increasing function
This is any sequence that is generated by \link{seq}
that increases in steps of one.
}
