% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conservation.R
\name{plot_conservation}
\alias{plot_conservation}
\title{Do the expected and observed number of SNPs
as a bar chart}
\usage{
plot_conservation(folder_name)
}
\arguments{
\item{folder_name}{name of the folder}
}
\description{
Do the expected and observed number of SNPs
as a bar chart
}
