% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conservation_per_spanner.R
\name{plot_conservation_per_spanner}
\alias{plot_conservation_per_spanner}
\title{Do the SNPs stats per spanner}
\usage{
plot_conservation_per_spanner(folder_name)
}
\arguments{
\item{folder_name}{name of the folder}
}
\description{
Do the SNPs stats per spanner
}
