% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_f_snps_found_and_expected.R
\name{plot_f_snps_found_and_expected}
\alias{plot_f_snps_found_and_expected}
\title{Plot}
\usage{
plot_f_snps_found_and_expected(folder_name)
}
\arguments{
\item{folder_name}{name of the folder}
}
\description{
Plot
}
