% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_f_snps_found_and_expected_per_n_tmhs.R
\name{plot_f_snps_found_and_expected_per_n_tmhs}
\alias{plot_f_snps_found_and_expected_per_n_tmhs}
\title{Plot}
\usage{
plot_f_snps_found_and_expected_per_n_tmhs(folder_name)
}
\description{
Plot
}
