% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_f_snps_found_and_expected_per_spanner.R
\name{plot_f_snps_found_and_expected_per_spanner}
\alias{plot_f_snps_found_and_expected_per_spanner}
\title{Plot}
\usage{
plot_f_snps_found_and_expected_per_spanner(folder_name)
}
\description{
Plot
}
