% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_f_tmh_ncbi_per_spanner.R
\name{plot_f_tmh_ncbi_per_spanner}
\alias{plot_f_tmh_ncbi_per_spanner}
\title{Plot the distribution of the percentage TMH that the NCBI proteins are}
\usage{
plot_f_tmh_ncbi_per_spanner(folder_name)
}
\description{
Plot the distribution of the percentage TMH that the NCBI proteins are
}
