% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_proteins_per_gene_name.R
\name{plot_n_proteins_per_gene_name}
\alias{plot_n_proteins_per_gene_name}
\title{The number of proteins per gene name}
\usage{
plot_n_proteins_per_gene_name(folder_name = folder_name)
}
\arguments{
\item{folder_name}{name of the folder}
}
\description{
The number of proteins per gene name
}
