% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_tmhs_per_protein.R
\name{plot_n_tmhs_per_protein}
\alias{plot_n_tmhs_per_protein}
\title{The number of TMHs per protein}
\usage{
plot_n_tmhs_per_protein(folder_name = folder_name)
}
\arguments{
\item{folder_name}{name of the folder}
}
\description{
The number of TMHs per protein
}
