% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snp_rel_pos.R
\name{plot_snp_rel_pos}
\alias{plot_snp_rel_pos}
\title{The relative positions of the SNPs}
\usage{
plot_snp_rel_pos(folder_name = folder_name)
}
\arguments{
\item{folder_name}{name of the folder}
}
\description{
The relative positions of the SNPs
}
