% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snps_per_gene_name.R
\name{plot_snps_per_gene_name}
\alias{plot_snps_per_gene_name}
\title{Create a plot that shows the number of SNPs per gene name}
\usage{
plot_snps_per_gene_name(folder_name = folder_name)
}
\arguments{
\item{folder_name}{name of the folder}
}
\description{
Create a plot that shows the number of SNPs per gene name
}
