% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snps_per_gene_name_processed.R
\name{plot_snps_per_gene_name_processed}
\alias{plot_snps_per_gene_name_processed}
\title{The number of SNPs processed}
\usage{
plot_snps_per_gene_name_processed(folder_name = folder_name)
}
\arguments{
\item{folder_name}{name of the folder}
}
\description{
The number of SNPs processed
}
