% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_f_snps_found_and_expected_per_spanner.R
\name{plot_f_snps_found_and_expected_per_spanner}
\alias{plot_f_snps_found_and_expected_per_spanner}
\title{Plot}
\usage{
plot_f_snps_found_and_expected_per_spanner(folder_name, use_color = TRUE)
}
\arguments{
\item{folder_name}{name of the folder}

\item{use_color}{set to TRUE to use color,
set to FALSE for black-and-white}
}
\description{
Plot
}
